% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_sentiment_analysis.R
\name{run_sentiment_analysis}
\alias{run_sentiment_analysis}
\title{Batch Sentiment Analysis for German Sentences
This function performs sentiment analysis on sentences written in German.}
\usage{
run_sentiment_analysis(sentences)
}
\arguments{
\item{sentences}{A character vector containing German sentences.}
}
\value{
A vector of sentiment scores.
}
\description{
Batch Sentiment Analysis for German Sentences
This function performs sentiment analysis on sentences written in German.
}
\examples{
sentences <- c("Das ist fantastisch!", "Das ist schrecklich.")
run_sentiment_analysis(sentences)
}
